+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/SharedIndexInformer.java
@@ -15,10 +15,12 @@
  */
 package io.fabric8.kubernetes.client.informers;
 
+import io.fabric8.kubernetes.client.WatcherException;
 import io.fabric8.kubernetes.client.informers.cache.Indexer;
 
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
 import java.util.function.Function;
 
 /**
@@ -39,4 +41,14 @@ public interface SharedIndexInformer<T> extends SharedInformer<T> {
    * @return the internal indexer store
    */
   Indexer<T> getIndexer();
+
+  /**
+   * Return a future that will allow notification of informer stopping.
+   * <p>
+   * If {@link #stop()} is called, the future will be completed with a null value.
+   * <p>
+   * If an exception occurs that terminates the informer, then it will be exceptionally completed with that exception
+   * - typically a {@link WatcherException}
+   */
+  CompletableFuture<Void> stopped();
 }