+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/impl/DefaultSharedIndexInformer.java
@@ -17,7 +17,6 @@ package io.fabric8.kubernetes.client.informers.impl;
 
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
-import io.fabric8.kubernetes.client.informers.InformerExceptionHandler;
 import io.fabric8.kubernetes.client.informers.ListerWatcher;
 import io.fabric8.kubernetes.client.informers.ResourceEventHandler;
 import io.fabric8.kubernetes.client.informers.ResyncRunnable;
@@ -35,6 +34,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
@@ -42,7 +42,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
-public class DefaultSharedIndexInformer<T extends HasMetadata, L extends KubernetesResourceList<T>> implements SharedIndexInformer<T> {
+public class DefaultSharedIndexInformer<T extends HasMetadata, L extends KubernetesResourceList<T>>
+    implements SharedIndexInformer<T> {
   private static final Logger log = LoggerFactory.getLogger(DefaultSharedIndexInformer.class);
 
   private static final long MINIMUM_RESYNC_PERIOD_MILLIS = 1000L;
@@ -68,9 +69,8 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
 
   private ScheduledFuture<?> resyncFuture;
 
-  private InformerExceptionHandler exceptionHandler = InformerExceptionHandler.NOOP;
-
-  public DefaultSharedIndexInformer(Class<T> apiTypeClass, ListerWatcher<T, L> listerWatcher, long resyncPeriod, Executor informerExecutor) {
+  public DefaultSharedIndexInformer(Class<T> apiTypeClass, ListerWatcher<T, L> listerWatcher, long resyncPeriod,
+      Executor informerExecutor) {
     if (resyncPeriod < 0) {
       throw new IllegalArgumentException("Invalid resync period provided, It should be a non-negative value");
     }
@@ -85,7 +85,7 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
     this.indexer.setIsRunning(this::isRunning);
 
     processorStore = new ProcessorStore<>(this.indexer, this.processor);
-    this.reflector = new Reflector<>(apiTypeClass, listerWatcher, processorStore, exceptionHandler);
+    this.reflector = new Reflector<>(apiTypeClass, listerWatcher, processorStore);
   }
 
   /**
@@ -107,13 +107,16 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
 
     if (resyncPeriodMillis > 0) {
       if (resyncPeriodMillis < MINIMUM_RESYNC_PERIOD_MILLIS) {
-        log.warn("DefaultSharedIndexInformer#resyncPeriod {} is too small. Changing it to minimal allowed value of {}", resyncPeriodMillis, MINIMUM_RESYNC_PERIOD_MILLIS);
+        log.warn("DefaultSharedIndexInformer#resyncPeriod {} is too small. Changing it to minimal allowed value of {}",
+            resyncPeriodMillis, MINIMUM_RESYNC_PERIOD_MILLIS);
         resyncPeriodMillis = MINIMUM_RESYNC_PERIOD_MILLIS;
       }
 
       if (resyncPeriodMillis < this.resyncCheckPeriodMillis) {
         if (started.get()) {
-          log.warn("DefaultSharedIndexInformer#resyncPeriod {} is smaller than resyncCheckPeriod {} and the informer has already started. Changing it to {}", resyncPeriodMillis, resyncCheckPeriodMillis,resyncCheckPeriodMillis);
+          log.warn(
+              "DefaultSharedIndexInformer#resyncPeriod {} is smaller than resyncCheckPeriod {} and the informer has already started. Changing it to {}",
+              resyncPeriodMillis, resyncCheckPeriodMillis, resyncCheckPeriodMillis);
           resyncPeriodMillis = resyncCheckPeriodMillis;
         } else {
           // if the event handler's resyncPeriod is smaller than the current resyncCheckPeriod
@@ -136,7 +139,7 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
   @Override
   public void run() {
     if (stopped) {
-        throw new IllegalStateException("Cannot restart a stopped informer");
+      throw new IllegalStateException("Cannot restart a stopped informer");
     }
     if (!started.compareAndSet(false, true)) {
       return;
@@ -146,7 +149,7 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
 
     scheduleResync(processor::shouldResync);
 
-    reflector.listSyncAndWatch();
+    reflector.start();
     // stop called while run is called could be ineffective, check for it afterwards
     synchronized (this) {
       if (stopped) {
@@ -179,7 +182,7 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
   public Indexer<T> getIndexer() {
     return this.indexer;
   }
-  
+
   @Override
   public Store<T> getStore() {
     return this.indexer;
@@ -200,17 +203,18 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
   public boolean isRunning() {
     return !stopped && started.get() && reflector.isRunning();
   }
-  
+
   @Override
   public boolean isWatching() {
     return reflector.isWatching();
   }
-  
+
   synchronized void scheduleResync(Supplier<Boolean> resyncFunc) {
     // schedule the resync runnable
     if (resyncCheckPeriodMillis > 0) {
       ResyncRunnable<T> resyncRunnable = new ResyncRunnable<>(processorStore, resyncFunc);
-      resyncFuture = Utils.scheduleAtFixedRate(informerExecutor, resyncRunnable, resyncCheckPeriodMillis, resyncCheckPeriodMillis, TimeUnit.MILLISECONDS);
+      resyncFuture = Utils.scheduleAtFixedRate(informerExecutor, resyncRunnable, resyncCheckPeriodMillis,
+          resyncCheckPeriodMillis, TimeUnit.MILLISECONDS);
     } else {
       log.debug("informer#Controller: resync skipped due to 0 full resync period {}", apiTypeClass);
     }
@@ -230,7 +234,8 @@ public class DefaultSharedIndexInformer<T extends HasMetadata, L extends Kuberne
   }
 
   @Override
-  public void setExceptionHandler(InformerExceptionHandler handler) {
-    this.exceptionHandler = handler != null ? handler : InformerExceptionHandler.NOOP;
+  public CompletableFuture<Void> stopped() {
+    return this.reflector.getStopFuture();
   }
+
 }